Reuse `reqwest::Client` when possible, and modular code
