+++ src/fetch.rs
@@ -1,53 +1,34 @@
 use anyhow::Context;
 use byteorder::{ByteOrder, LittleEndian};
 use reqwest::{
-    header::{HeaderMap, HeaderValue, AUTHORIZATION, RANGE},
+    header::{HeaderValue, RANGE},
     Client, StatusCode,
 };
-use serde::Deserialize;
 use std::collections::HashMap;
+
+use crate::schemas::FileType;
 
 // As per https://github.com/huggingface/huggingface_hub/pull/1855#discussion_r1404286419, we
 // only fetch the first 100kb of the `model.safetensors` file, as empirically, 97% of
 // safetensors files have a metadata size < 100kb (over the top 1000 models on the Hub)
 static MAX_METADATA_SIZE: usize = 100_000;
 
-#[derive(Deserialize, Debug)]
-pub struct FileType {
-    pub dtype: Option<String>,
-    pub shape: Option<Vec<u64>>,
-    data_offsets: Option<(u64, u64)>,
-}
-
-// TODO: we need to first check if the files are available in the Hugging Face cache, if so, just read those directly; if not, then fetch those from the Hub (probably after the first release)
 async fn fetch_metadata(
+    client: &Client,
     url: &str,
-    token: &str,
     range_lower: Option<usize>,
     range_upper: Option<usize>,
 ) -> anyhow::Result<(bytes::Bytes, usize)> {
-    let mut headers = HeaderMap::new();
-    headers.insert(
-        AUTHORIZATION,
-        HeaderValue::from_str(format!("Bearer {}", token).as_str())
-            .context("parsing the authorization header with the hf token failed")?,
-    );
     let range_lower = range_lower.unwrap_or(0);
     let range_upper = range_upper.unwrap_or(MAX_METADATA_SIZE);
-    headers.insert(
-        RANGE,
-        HeaderValue::from_str(format!("bytes={}-{}", range_lower, range_upper).as_str())
-            .context("parsing the range of bytes to fetch for the range header failed")?,
-    );
-
-    // TODO: do we really need a client?
-    let client = Client::builder()
-        .default_headers(headers)
-        .build()
-        .context("couldn't build the reqwest client")?;
 
     match client
         .get(url)
+        .header(
+            RANGE,
+            HeaderValue::from_str(format!("bytes={}-{}", range_lower, range_upper).as_str())
+                .context("parsing the range of bytes to fetch for the range header failed")?,
+        )
         .send()
         .await
         .context("fetching the provided url failed")
@@ -67,13 +48,13 @@ async fn fetch_metadata(
     }
 }
 
-pub async fn fetch(url: &str, token: &str) -> anyhow::Result<HashMap<String, FileType>> {
-    let (metadata, metadata_size) = fetch_metadata(url, token, None, None)
+pub async fn fetch(client: &Client, url: &str) -> anyhow::Result<HashMap<String, FileType>> {
+    let (metadata, metadata_size) = fetch_metadata(&client, &url, None, None)
         .await
         .context("failed fetching the metadata from the provided url")?;
 
     let metadata = if metadata_size > MAX_METADATA_SIZE {
-        fetch_metadata(url, token, Some(8), Some(metadata_size))
+        fetch_metadata(&client, &url, Some(8), Some(metadata_size))
             .await
             .context("failed fetching the complete metadata from the provided url")?
             .0
@@ -81,8 +62,6 @@ pub async fn fetch(url: &str, token: &str) -> anyhow::Result<HashMap<String, Fil
         metadata
     };
 
-    // TODO: maybe just a Vec<FileType> with the values is more efficient since the keys are not
-    // used?
     serde_json::from_slice::<HashMap<String, FileType>>(&metadata[8..metadata_size + 8])
         .context("parsing the metadata bytes into a serde_json::Value failed")
 }
